+++ CHANGELOG.md
@@ -21,8 +21,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `--debug` didn't work because of a tiny error with `modules/log.py` (1.0.0 -> 1.0.1)
 - `genscript.py`: Some cleanup
+- `src.py 1.0.1`: Various cleanup and fixes.
 - `config.py 1.0.1`: Various cleanup and fixes.
 - `makezip.py 1.0.2`: Various cleanup and fixes.
+- `theme.py 1.0.1`: Various cleanup. More error handling.
 - `exclusion.py 1.0.1`: Various cleanup. More error handling.
 - `inclusion.py 1.0.1`: Various cleanup. More error handling.